Refactor App.js to use separate screens for Start and Confirm, and update related components and styling
